\name{file.ext}
\alias{file.ext}
\title{file.ext}
\usage{
  file.ext(x, sep = ".", lowercase = TRUE)
}
\arguments{
  \item{x}{file name or path}

  \item{sep}{file extension separator}

  \item{lowercase}{convert return value to lowercase
  first?}
}
\value{
  the file extension, such as .jpg or .txt
}
\description{
  Given a filename, returns the file extension
}

