\name{ggplot.zoo}
\alias{ggplot.zoo}
\title{ggplot.zoo}
\usage{
  ggplot.zoo(data, ...)
}
\arguments{
  \item{data}{a \link[zoo]{zoo} object}

  \item{...}{further arguments to \link[ggplot2]{ggplot}}
}
\value{
  same thing as \link[ggplot2]{ggplot}
}
\description{
  Plot a zoo object using ggplot2
}

