\name{update.list}
\alias{update.list}
\title{update.list}
\usage{
  update.list(object, ...)
}
\arguments{
  \item{object}{a list}

  \item{...}{arguments of the form key=value}
}
\value{
  a copy of the list with updated items
}
\description{
  Update the value of items in a list (possibly adding new
  items).
}

